Move `NodeRelation` to dbt/artifacts
